requirements.txt
Django==5.1
python-jobspy==1.1.80
linkedin-api==2.3.1
openai==1.93.0
pandas==2.3.0
mistune
dotenv